+++ modules/apps/29-fee/ibc_middleware_test.go
@@ -525,7 +525,7 @@ func (suite *FeeTestSuite) TestOnRecvPacket() {
 				forwardAddr, _ := suite.chainB.GetSimApp().IBCFeeKeeper.GetCounterpartyAddress(suite.chainB.GetContext(), suite.chainA.SenderAccount.GetAddress().String(), suite.path.EndpointB.ChannelID)
 
 				expectedAck := types.IncentivizedAcknowledgement{
-					Result:                ibcmock.MockAcknowledgement.Acknowledgement(),
+					AppAcknowledgement:    ibcmock.MockAcknowledgement.Acknowledgement(),
 					ForwardRelayerAddress: forwardAddr,
 					UnderlyingAppSuccess:  true,
 				}
@@ -544,7 +544,7 @@ func (suite *FeeTestSuite) TestOnRecvPacket() {
 
 			case !tc.forwardRelayer:
 				expectedAck := types.IncentivizedAcknowledgement{
-					Result:                ibcmock.MockAcknowledgement.Acknowledgement(),
+					AppAcknowledgement:    ibcmock.MockAcknowledgement.Acknowledgement(),
 					ForwardRelayerAddress: "",
 					UnderlyingAppSuccess:  true,
 				}
@@ -583,7 +583,7 @@ func (suite *FeeTestSuite) TestOnAcknowledgementPacket() {
 				suite.chainA.GetSimApp().IBCFeeKeeper.DeleteFeesInEscrow(suite.chainA.GetContext(), packetID)
 
 				ack = types.IncentivizedAcknowledgement{
-					Result:                ibcmock.MockAcknowledgement.Acknowledgement(),
+					AppAcknowledgement:    ibcmock.MockAcknowledgement.Acknowledgement(),
 					ForwardRelayerAddress: suite.chainA.SenderAccount.GetAddress().String(),
 				}.Acknowledgement()
 
@@ -625,7 +625,7 @@ func (suite *FeeTestSuite) TestOnAcknowledgementPacket() {
 				blockedAddr := suite.chainA.GetSimApp().AccountKeeper.GetModuleAccount(suite.chainA.GetContext(), transfertypes.ModuleName).GetAddress()
 
 				ack = types.IncentivizedAcknowledgement{
-					Result:                ibcmock.MockAcknowledgement.Acknowledgement(),
+					AppAcknowledgement:    ibcmock.MockAcknowledgement.Acknowledgement(),
 					ForwardRelayerAddress: blockedAddr.String(),
 				}.Acknowledgement()
 
@@ -672,7 +672,7 @@ func (suite *FeeTestSuite) TestOnAcknowledgementPacket() {
 
 			// must be changed explicitly
 			ack = types.IncentivizedAcknowledgement{
-				Result:                ibcmock.MockAcknowledgement.Acknowledgement(),
+				AppAcknowledgement:    ibcmock.MockAcknowledgement.Acknowledgement(),
 				ForwardRelayerAddress: relayerAddr.String(),
 			}.Acknowledgement()